PL-206 - DAG IDs should be int64_t
Summary: All our IDs are int64_t.

Test Plan: Modified unit tests.

Reviewers: #engineering, michelle, kgandhi

Reviewed By: #engineering, kgandhi

Differential Revision: https://phab.corp.pixielabs.ai/D179

GitOrigin-RevId: a1442e63cb34ae8f15fac55a04c01c6dbe50c661

+++ src/carnot/plan/dag.h
@@ -20,39 +20,39 @@ namespace plan {
 
 class DAG {
  public:
-  void AddNode(int node);
-  void DeleteNode(int node);
+  void AddNode(int64_t node);
+  void DeleteNode(int64_t node);
 
-  bool HasNode(int node);
+  bool HasNode(int64_t node);
 
-  void AddEdge(int from_node, int to_node);
-  void DeleteEdge(int from_node, int to_node);
+  void AddEdge(int64_t from_node, int64_t to_node);
+  void DeleteEdge(int64_t from_node, int64_t to_node);
   std::string DebugString();
   void Debug();
 
-  std::unordered_set<int> Orphans();
-  std::unordered_set<int> TransitiveDepsFrom(int node);
-  std::vector<int> TopologicalSort();
+  std::unordered_set<int64_t> Orphans();
+  std::unordered_set<int64_t> TransitiveDepsFrom(int64_t node);
+  std::vector<int64_t> TopologicalSort();
 
-  std::vector<int> DependenciesOf(int node) {
+  std::vector<int64_t> DependenciesOf(int64_t node) {
     if (nodes_.find(node) != std::end(nodes_)) {
       return forward_edges_by_node_[node];
     }
     return {};
   }
 
-  const std::unordered_set<int>& nodes() { return nodes_; }
+  const std::unordered_set<int64_t>& nodes() { return nodes_; }
 
  private:
   // Store all the integer id's as nodes.
-  std::unordered_set<int> nodes_;
+  std::unordered_set<int64_t> nodes_;
 
   // This is much more efficient to do in a fstar/rstar structure, but the dual adjacency
   // list is much simpler to update and this is likely not on the critical path.
   // These are the orderered edges from this node to others.
-  std::unordered_map<int, std::vector<int>> forward_edges_by_node_;
+  std::unordered_map<int64_t, std::vector<int64_t>> forward_edges_by_node_;
   // These are ids of the predecesors nodes.
-  std::unordered_map<int, std::vector<int>> reverse_edges_by_node_;
+  std::unordered_map<int64_t, std::vector<int64_t>> reverse_edges_by_node_;
 };
 
 }  // namespace plan
